finish backend setup

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 
 const asyncHandler = require('express-async-handler')
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -19,7 +20,8 @@ const setGoals = asyncHandler(async(req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -31,10 +33,25 @@ const setGoals = asyncHandler(async(req, res) => {
 const updateGoals = asyncHandler(async(req, res) => {
     const goal = await Goal.findById(req.params.id)
 
+
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.status(200).json(updatedGoal)
@@ -48,6 +65,18 @@ const deleteGoals = asyncHandler(async(req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     await goal.deleteOne()
